Move profile estimation into new formulas
