+++ cfspopcon/algorithms.yaml
@@ -1,5 +1,24 @@
 # Autogenerated by Algorithm.write_yaml()
 
+calc_1D_plasma_profiles:
+  inputs:
+  - profile_form
+  - average_electron_density
+  - average_electron_temp
+  - average_ion_temp
+  - electron_density_peaking
+  - ion_density_peaking
+  - temperature_peaking
+  - dilution
+  optionals:
+  - normalized_inverse_temp_scale_length
+  - npoints
+  returns:
+  - rho
+  - electron_density_profile
+  - fuel_ion_density_profile
+  - electron_temp_profile
+  - ion_temp_profile
 calc_P_SOL:
   inputs:
   - P_in
@@ -144,6 +163,25 @@ calc_edge_impurity_concentration:
   - edge_impurity_concentration
   - edge_impurity_concentration_in_core
   - impurities
+calc_effective_collisionality:
+  inputs:
+  - average_electron_density
+  - average_electron_temp
+  - major_radius
+  - z_effective
+  optionals: []
+  returns:
+  - effective_collisionality
+calc_electron_density_peaking:
+  inputs:
+  - effective_collisionality
+  - beta_toroidal
+  - electron_density_peaking_offset
+  - average_electron_density
+  optionals: []
+  returns:
+  - electron_density_peaking
+  - peak_electron_density
 calc_extrinsic_core_radiator:
   inputs:
   - minimum_core_radiated_fraction
@@ -241,6 +279,17 @@ calc_heat_exhaust:
   - lambda_q
   - q_parallel
   - q_perp
+calc_ion_density_peaking:
+  inputs:
+  - effective_collisionality
+  - beta_toroidal
+  - ion_density_peaking_offset
+  - average_electron_density
+  - dilution
+  optionals: []
+  returns:
+  - ion_density_peaking
+  - peak_fuel_ion_density
 calc_magnetic_field_on_axis:
   inputs:
   - product_of_magnetic_field_and_radius
@@ -398,6 +447,15 @@ calc_separatrix_triangularity_from_triangularity95:
   optionals: []
   returns:
   - separatrix_triangularity
+calc_temperature_peaking:
+  inputs:
+  - average_electron_temp
+  - average_ion_temp
+  - temperature_peaking
+  optionals: []
+  returns:
+  - peak_electron_temp
+  - peak_ion_temp
 calc_triple_product:
   inputs:
   - peak_fuel_ion_density
