--- cfspopcon/deprecated_formulas/plasma_profile_data/__init__.py
@@ -1,1 +1,0 @@
-"""1D plasma profiles based on TRANSP runs."""
