+++ cfspopcon/formulas/plasma_profiles/__init__.py
@@ -1,1 +1,5 @@
 """Routines to calculate estimate 1D profiles for the confined region."""
+
+from . import density_peaking, plasma_profiles, temperature_peaking
+
+__all__ = ["density_peaking", "temperature_peaking", "plasma_profiles"]
